Fix issues

+++ modules/yup_audio_devices/native/yup_CoreAudio_mac.cpp
@@ -1158,7 +1158,7 @@ struct CoreAudioClasses
         {
             auto& intern = *static_cast<CoreAudioInternal*> (inClientData);
 
-            const auto xruns = std::count_if (pa, pa + numAddresses, [] (const AudioObjectPropertyAddress& x)
+            const auto xruns = (int) std::count_if (pa, pa + numAddresses, [] (const AudioObjectPropertyAddress& x)
             {
                 return x.mSelector == kAudioDeviceProcessorOverload;
             });
